Introduce routing with page.js
This included making changes that would have routing make sense. And
changes that became necessary on account of having multiple pages in
this SPA. This introduces a list of characters to the reducer, it moves
all events off into a list of objects that are returned from a closure
with a reference to the store. And of course there is now a second page
which lists all of the *very* simple "characters" that can be made.