Add Dragula for drag-and-drop
There's still more than a bit to be worked out here. Right now the
drag-and-drop that I have is really not so great. You need to shake the
thing being dragged all over the screen to get it to work right, I'm not
sure if it has anything to do with the type of content I'm dragging? If
maybe div's work better than li's, or visa-versa? Have to do more
exploration on that.